app/register/x -> app/users/register/x
